Reflect code review comments
JIRA: ONE-4838
